+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * Update to `col_hide(columns=)` and `cols_unhide(columns=)` arguments. The functions no longer return an error if no columns are selected. (#1342)
 
+* Update word processing to htmlEscape all characters before it goes into xml_t. This means now footnotes are escaped too. (#1303)
+
 # gt 0.9.0
 
 ## New features